requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot
python-dotenv
langgraph
langchain-core
google-generativeai
requests
beautifulsoup4
gnews
yfinance
pandas
pandas-ta
mplfinance
numpy
scipy
reportlab
matplotlib